working on adding more resources and will be working on inventory to allow more functionality.

+++ Game/src/core/GroundItem.java
@@ -5,7 +5,7 @@ import ui.UITextureType;
 public class GroundItem extends Tile {
 	public UITextureType type = UITextureType.BLANK;
 	public ItemType item = ItemType.NONE;
-	public int count = 0;
+	public long count = 0;
 
 	public GroundItem(TextureType newType) {
 		super(newType);
